% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_summary_class}
\alias{.add_summary_class}
\title{Add S3 class to a summary object.}
\usage{
.add_summary_class(summary, summary_class)
}
\arguments{
\item{summary}{\code{list} - contains the summary.}

\item{summary_class}{\code{character} - the summary object will have the class
c("summary_<summary_class>", "summary").}
}
\value{
\code{summary} - summary object.
}
\description{
Add S3 class to a summary object.
}
\keyword{internal}
