% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assert_subset}
\alias{.assert_subset}
\title{Assert that required values are a subset of available values}
\usage{
.assert_subset(required_cols, available_cols, dataset_name, col_type = NULL)
}
\arguments{
\item{required_cols}{\code{character} - values to check for.}

\item{available_cols}{\code{character} - available values (e.g., from
colnames()).}

\item{dataset_name}{\code{character} - name of the data object for error
message.}

\item{col_type}{\code{character} - optional descriptor for value type
(e.g., "ID", "Flag").}
}
\value{
\code{NULL} if assertion passes, otherwise throws an error.
}
\description{
Assert that required values are a subset of available values
}
\keyword{internal}
