% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_columns}
\alias{.format_columns}
\title{Format columns from 'col1_col2' to 'col1, col2'}
\usage{
.format_columns(columns)
}
\arguments{
\item{columns}{\code{character} - columns to format.}
}
\value{
\code{character} - formatted columns.
}
\description{
Format columns from 'col1_col2' to 'col1, col2'
}
\keyword{internal}
