% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_utils.R
\name{.get_unique_values}
\alias{.get_unique_values}
\title{Glimpse a vector creating a summary of unique values.}
\usage{
.get_unique_values(x, col_name)
}
\arguments{
\item{x}{\code{vector} - input vector.}

\item{col_name}{\code{character} - column name corresponding to the input vector.}
}
\value{
\code{summary} - contains unique values.
}
\description{
Glimpse a vector creating a summary of unique values.
}
\keyword{internal}
