% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_utils.R
\name{.inject_nas}
\alias{.inject_nas}
\title{Inject NAs into a vector.}
\usage{
.inject_nas(x, na_positions)
}
\arguments{
\item{x}{\code{vector} - simulated vector.}

\item{na_positions}{\code{integer} - indices where NAs should be injected.}
}
\value{
\code{vector} - with NAs injected.
}
\description{
Inject NAs into a vector.
}
\keyword{internal}
