% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_study.R
\name{.simulate_from_summary}
\alias{.simulate_from_summary}
\title{Dispatch a single summary to its corresponding simulate_* function.}
\usage{
.simulate_from_summary(summary, seed)
}
\arguments{
\item{summary}{classed summary object (\code{summary_adsl} / \code{summary_bds} /
\code{summary_tte} / \code{summary_occds}).}

\item{seed}{\code{integer} or \code{NULL} - simulation seed.}
}
\value{
\code{data.frame} of synthetic data.
}
\description{
Dispatch a single summary to its corresponding simulate_* function.
}
\keyword{internal}
