% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_adsl.R
\name{.validate_input_glimpse_adsl}
\alias{.validate_input_glimpse_adsl}
\title{Validate input arguments for \code{glimpse_adsl()}.}
\usage{
.validate_input_glimpse_adsl(
  adsl,
  id_cols,
  treatment_cols,
  flag_cols,
  ordered_col_sets
)
}
\arguments{
\item{adsl}{\code{data.frame} - input ADSL dataset.}

\item{id_cols}{\code{character} - ID columns e.g. "USUBJID".}

\item{treatment_cols}{\code{character} - treatment columns e.g. "TRT01A".}

\item{flag_cols}{\code{character} - flag columns e.g. "SAFFL". Default is an
empty character vector (no flag columns).}

\item{ordered_col_sets}{\verb{list<character>} - each element is a character
vector naming columns whose combinations should be preserved e.g.
\code{list(c("REGION1", "REGION1N"))}. Default is an empty list.}
}
\value{
None.
}
\description{
Validate input arguments for \code{glimpse_adsl()}.
}
\keyword{internal}
