% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_bds.R
\name{.validate_input_glimpse_bds}
\alias{.validate_input_glimpse_bds}
\title{Validate input arguments for \code{glimpse_bds()}.}
\usage{
.validate_input_glimpse_bds(
  bds,
  syn_adsl,
  id_cols,
  param_cols,
  visit_cols,
  flag_cols,
  ordered_col_sets
)
}
\arguments{
\item{bds}{\code{data.frame} - input BDS dataset e.g. ADLB.}

\item{syn_adsl}{\code{data.frame} - synthetic ADSL to obtain subject-level data
from. Should be created by \code{\link[=simulate_adsl]{simulate_adsl()}}.}

\item{id_cols}{\code{character} - ID column names that identify subjects e.g.
c("USUBJID"). These columns must exist in \code{syn_adsl} and will be taken
from there (never simulated).}

\item{param_cols}{\code{character} - parameter columns e.g. c("PARAM", "PARAMCD").}

\item{visit_cols}{\code{character} - visit columns e.g. c("AVISIT", "AVISITN").
Default is an empty character vector (no visit columns).}

\item{flag_cols}{\code{character} - flag columns that will be sampled as Y/N in
proportion to input e.g. c("ANL01FL"). Default is an empty character vector
(no flag columns).}

\item{ordered_col_sets}{\verb{list<character>} - each element is a character
vector naming columns whose combinations should be preserved e.g.
\code{list(c("REGION1", "REGION1N"))}. Default is an empty list.}
}
\value{
None.
}
\description{
Validate input arguments for \code{glimpse_bds()}.
}
\keyword{internal}
