% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_occds.R
\name{.validate_input_glimpse_occds}
\alias{.validate_input_glimpse_occds}
\title{Validate input arguments for \code{glimpse_occds()}.}
\usage{
.validate_input_glimpse_occds(
  occds,
  syn_adsl,
  id_cols,
  seq_col,
  flag_cols,
  ordered_col_sets
)
}
\arguments{
\item{occds}{\code{data.frame} - input OCCDS dataset e.g. ADAE, ADCM.}

\item{syn_adsl}{\code{data.frame} - synthetic ADSL to obtain subject-level data
from. Should be created by \code{\link[=simulate_adsl]{simulate_adsl()}}.}

\item{id_cols}{\code{character} - ID column names that identify subjects e.g.
c("USUBJID"). These columns must exist in \code{syn_adsl} and will be taken
from there (never simulated).}

\item{seq_col}{\code{character} - sequence column name e.g. "AESEQ", "CMSEQ".}

\item{flag_cols}{\code{character} - flag columns that will be sampled as Y/N in
proportion to input e.g. c("ANL01FL"). Default is an empty character vector
(no flag columns).}

\item{ordered_col_sets}{\verb{list<character>} - each element is a character
vector naming columns whose combinations should be preserved e.g.
\code{list(c("REGION1", "REGION1N"))}. Default is an empty list.}
}
\value{
None.
}
\description{
Validate input arguments for \code{glimpse_occds()}.
}
\keyword{internal}
