% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_tte.R
\name{.validate_input_glimpse_tte}
\alias{.validate_input_glimpse_tte}
\title{Validate input arguments for \code{glimpse_tte()}.}
\usage{
.validate_input_glimpse_tte(
  tte,
  syn_adsl,
  param_cols,
  censor_cols,
  flag_cols,
  ordered_col_sets
)
}
\arguments{
\item{tte}{\code{data.frame} - input TTE dataset e.g. ADTTE.}

\item{syn_adsl}{\code{data.frame} - synthetic ADSL to obtain subject-level data
from. Should be created by \code{\link[=simulate_adsl]{simulate_adsl()}}.}

\item{param_cols}{\code{character} - parameter columns e.g. c("PARAM", "PARAMCD").}

\item{censor_cols}{\code{character} - censoring columns e.g.
c("CNSR", "EVNTDESC", "CNSDTDSC"). These columns are treated as an ordered
set to preserve co-occurrence patterns. Only needed when multiple censor
columns should be kept together. Default is an empty character vector
(censor columns are treated as regular columns).}

\item{flag_cols}{\code{character} - flag columns that will be sampled as Y/N in
proportion to input e.g. c("ANL01FL"). Default is an empty character vector
(no flag columns).}

\item{ordered_col_sets}{\verb{list<character>} - each element is a character
vector naming columns whose combinations should be preserved e.g.
\code{list(c("REGION1", "REGION1N"))}. Default is an empty list.}
}
\value{
None.
}
\description{
Validate input arguments for \code{glimpse_tte()}.
}
\keyword{internal}
