% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_study.R
\name{.validate_study_summary}
\alias{.validate_study_summary}
\title{Validate the structure of a study summary.}
\usage{
.validate_study_summary(study_summary)
}
\arguments{
\item{study_summary}{\code{list} - object read from a glimpse_study() rds.}
}
\value{
None. Throws on malformed input.
}
\description{
Validate the structure of a study summary.
}
\keyword{internal}
