% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_study_config.R
\name{.yaml_quote}
\alias{.yaml_quote}
\title{Wrap a string in double quotes, escaping any embedded double quotes.}
\usage{
.yaml_quote(s)
}
\arguments{
\item{s}{\code{character(1)} - string to quote.}
}
\value{
\code{character(1)} - the YAML-quoted string.
}
\description{
Wrap a string in double quotes, escaping any embedded double quotes.
}
\keyword{internal}
