% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_vec.R
\name{glimpse_id}
\alias{glimpse_id}
\title{Glimpse an ID vector.}
\usage{
glimpse_id(x, col_name)
}
\arguments{
\item{x}{\code{vector} - contains ID values.}

\item{col_name}{\code{character} - column name corresponding to the input vector.}
}
\value{
\code{summary} - contains column name for ID generation.
}
\description{
Glimpse an ID vector.
}
\examples{
# Glimpse an ID column, then simulate fresh sequential IDs.
summary <- glimpse_id(c("S1", "S2", "S3"), col_name = "USUBJID")
simulate(summary, output_length = 5)

}
