% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_sas7bdat}
\alias{read_sas7bdat}
\title{Read in an ADaM dataset.}
\usage{
read_sas7bdat(adam_path)
}
\arguments{
\item{adam_path}{\code{character} - path to the \code{.sas7bdat} file containing the
ADaM dataset.}
}
\value{
\code{data.frame} - ADaM dataset.
}
\description{
Loads in a ADaM dataset and and convert all missing values to NA.
}
\examples{
\dontshow{
# Write a small SAS dataset to a temporary file.
adsl <- read.csv(
  system.file("extdata", "adsl.csv", package = "synadam"),
  stringsAsFactors = FALSE
)
sas_path <- tempfile(fileext = ".sas7bdat")
suppressWarnings(haven::write_sas(adsl, sas_path))
}
adam <- read_sas7bdat(sas_path)
head(adam)

}
