% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_df.R, R/simulate_vec.R
\name{simulate.summary_ordered}
\alias{simulate.summary_ordered}
\alias{simulate.summary_treatment_flag}
\alias{simulate}
\alias{simulate.summary_character}
\alias{simulate.summary_Date}
\alias{simulate.summary_POSIXct}
\alias{simulate.summary_difftime}
\alias{simulate.summary_integer}
\alias{simulate.summary_double}
\alias{simulate.summary_id}
\alias{simulate.summary_flag}
\title{Simulate a vector from a summary object.}
\usage{
\method{simulate}{summary_ordered}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_treatment_flag}(summary, ...)

simulate(summary, ...)

\method{simulate}{summary_character}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_Date}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_POSIXct}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_difftime}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_integer}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_double}(summary, output_length, seed = NULL, ...)

\method{simulate}{summary_id}(summary, output_length, ...)

\method{simulate}{summary_flag}(summary, output_length, seed = NULL, ...)
}
\arguments{
\item{summary}{\code{summary} - output from a glimpse function.}

\item{output_length}{\code{integer} - number of rows to simulate.}

\item{seed}{\code{integer} - random seed for reproducibility.}

\item{...}{additional arguments passed to methods.}
}
\value{
\code{data.frame} - the simulated dataset.
}
\description{
Simulate a vector from a summary object.
}
\section{Methods (by class)}{
\itemize{
\item \code{simulate(summary_ordered)}: simulate ordered columns by sampling from unique
combinations.

\item \code{simulate(summary_treatment_flag)}: simulate treatment and flag columns by preserving
exact counts for each combination.

\item \code{simulate(summary_character)}: simulate a character vector by sampling from unique
values.

\item \code{simulate(summary_Date)}: simulate a Date vector by sampling uniformly
between min and max.

\item \code{simulate(summary_POSIXct)}: simulate a POSIXct vector by sampling uniformly
between min and max.

\item \code{simulate(summary_difftime)}: simulate a difftime vector by sampling uniformly
between min and max.

\item \code{simulate(summary_integer)}: simulate an integerish vector by sampling uniformly
between min and max and rounding.

\item \code{simulate(summary_double)}: simulate a double vector by sampling uniformly
between min and max.

\item \code{simulate(summary_id)}: simulate an ID vector by generating sequential IDs.

\item \code{simulate(summary_flag)}: simulate a flag vector by sampling Y/N in proportion
to input.

}}
\examples{
# Summarise a vector with glimpse().
summary <- glimpse(c("A", "B", "B", "C"), col_name = "column_name")
simulate(summary, output_length = 10, seed = 1)

}
